client/requirements.txt
Django==3.2.15
grpcio==1.39.0
grpcio-tools==1.39.0
django-grpc==1.0.13

